backend/requirements.txt
# Cyprus For Christ - Backend Requirements
# Python 3.8+

# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0

# MySQL - Primary Database
mysqlclient>=2.2.0
django-mysql>=4.12.0
mysql-connector-python>=8.2.0
PyMySQL>=1.1.0

# Authentication & Security
django-otp>=1.2.0
qrcode>=7.4.0
pyotp>=2.8.0
cryptography>=41.0.0

# OpenAI Integration
openai
google-generativeai>=1.0.0

# PayPal Integration
paypalrestsdk>=1.13.0

# Utilities
python-decouple>=3.8
requests>=2.31.0
Pillow>=10.0.0
python-slugify>=8.0.0
python-dateutil>=2.8.0

# PDF Generation (for receipts)
reportlab>=4.0.0
pypdf2>=3.0.0

# API Documentation
drf-yasg>=1.21.0

# Development
django-debug-toolbar>=4.0.0
